move to dedicated class
